01 - This code has the implementation of a Rude AI
in this code I try and understand the Importance and influence o a proper prompt template, I constructed a prompt instructing the model to reply rudely to all the responses, it does a great job and is kinda funny tbh. fun way to understand the basic concepts and building a simple GenAI POC app. 

Install requirements.txt using pip install -r requirements.txt.