Added code for the basic rag Implementation
In this file we explore the implementation of a RAG(Retrieval Augmented System) using the OpenAI api. the default code uses OpenAI Embeddings & model rather than llama2, but the code for using that is given in comments. just a couple of lines change, embeddings and the model, the chain, db, prompt everything stays same. be sure to import the libraries though and obviously have llama2 installed using Ollama in your system. we read a pdf file here, but code for reading web pages is also given in comments, for further usage check official documentation🙂.