Comprehensive Report - using Structured Tools
Structured tools help keep code more organised, readable, and extendable. In this code I implemented a structured tool that provides a comprehensive Topic report on a given topic, uses newsapi for getting the news, ChatOpenAI for getting facts about the topic, and Textblob for easily analysing the sentiment of the topic based on description of articles provided by the news api. If you're using this.
1. Install textblob to the pre-existing requirements.txt
2. get the news_api key from ' https://newsapi.org/ '
and set them in the code